requirements.txt
larousse-api-sunbro==0.0.5          # Definitions from Larousse